requirements.txt
dotenv
faiss-cpu 
PyPDF2 
sentence-transformers 
google-generativeai